# Optimizing the go-parser