--- a/.editorconfig
+++ b/.editorconfig
@@ -1,0 +1,4 @@
+[*]
+trim_trailing_whitespace = true
+insert_final_newline = true
+