--- a/src/parser.rs
+++ b/src/parser.rs
@@ -52,23 +52,21 @@ struct ProjectRoot {
 fn find_project_roots(root_path: impl AsRef<Path>) -> Vec<ProjectRoot> {
     let mut project_roots = Vec::new();
 
-    for result in Walk::new(root_path) {
-        if let Ok(entry) = result {
-            if !entry.file_type().map_or(false, |ft| ft.is_file()) {
-                continue;
-            }
+    for entry in Walk::new(root_path).flatten() {
+        if !entry.file_type().map_or(false, |ft| ft.is_file()) {
+            continue;
+        }
 
-            let path = entry.path();
-            let file_name = path.file_name().and_then(|n| n.to_str()).unwrap_or("");
-            let parent_path = path.parent();
+        let path = entry.path();
+        let file_name = path.file_name().and_then(|n| n.to_str()).unwrap_or("");
+        let parent_path = path.parent();
 
-            if let Some(parent) = parent_path {
-                if let Some(project_type) = Language::from_file_name(file_name) {
-                    project_roots.push(ProjectRoot {
-                        path: parent.to_path_buf(),
-                        project_type,
-                    });
-                }
+        if let Some(parent) = parent_path {
+            if let Some(project_type) = Language::from_file_name(file_name) {
+                project_roots.push(ProjectRoot {
+                    path: parent.to_path_buf(),
+                    project_type,
+                });
             }
         }
     }
@@ -101,13 +99,13 @@ pub fn parse_root(root_path: impl AsRef<Path>, language: Option<&str>) -> Vec<Li
 }
 
 fn matches_language(project_type: Language, language: &str) -> bool {
-    match (project_type, language.to_lowercase().as_str()) {
-        (Language::Rust(_), "rust") => true,
-        (Language::Node(_), "node") => true,
-        (Language::Go(_), "go") => true,
-        (Language::Python(_), "python") => true,
-        _ => false,
-    }
+    matches!(
+        (project_type, language.to_lowercase().as_str()),
+        (Language::Rust(_), "rust")
+            | (Language::Node(_), "node")
+            | (Language::Go(_), "go")
+            | (Language::Python(_), "python")
+    )
 }
 
 fn parse_dependencies(root: &ProjectRoot) -> Vec<LicenseInfo> {