* Improves the time to parse go projects:

```
from:
cargo run -- -p ../go-ethereum  9.45s user 1.13s system 9% cpu 1:52.8 total
to:
cargo run -- -p ../go-ethereum  4.01s user 0.46s system 29% cpu 15.126 total
```

* Adds clippy warning for better linting